Call an external program for HTML stripping
If -S is given, the argument will be a program that will read HTML from
stdin and it will output a stripped version of it.

The program is executed by the shell so we allow the user to execute a
piped command for example.

-S implies -s and if both are given, -S will have priority over -s.

The input/output is stored as bytes and converter to UTF-8 by us.