Filter duplicated posts based on their content.
Some feeds like Youtube update their posts without changing their
content.

Filtering them based on the update date is useless and filtering
them by link is problematic because a post could be updated for good
reasons like for amending.

The solution is use a hash of the post's content as a key to filter out
duplicated posts.

This commit implements the filtering but it does not persists the hashes
in the database yet.