Rename db to dbfile; set a valid default for dbdata
Minor refactor renaming self.db to self.dbfile which makes explicit that
self.dbfile it is a file name and not the database loaded in memory.

If the database does not exist or it is invalid, self.dbdata was set to
None. Instead, set it to an empty dictionary. This makes clear the
format of an empty database.