Refactor run() and find_new()
Move the for-loop of run() into a separated method: compose_emails()

Move the update of the database from find_new() into a separated method:
update_database(); and ensure that it is called at the end of run()

Originally find_new() allowed to write/update the database conditionally
and even in an alternative file but nobody is using these
customizations.

update_database() allows to set an alternative file but the update is
unconditional. If needed, run() can be modified to make the update of
the database optional.