New DB scheme: store posts_seen and move feeds in a sub key
The original DB scheme (named 0) had the feeds' names as key of a
dictionary and the last updated timestamps as their values.

In the new DB scheme (named 1) that data is in a subdictionary index by
the 'feeds_last_time_checked' key.

Next to that two more keys: 'posts_seen' and 'feed2maildir_db_scheme'.

The latter just says which DB scheme is in use. The former is a
dictionary which keys are posts' hashes and values a timestamp when the
posts were seen.

The DB scheme change takes into consideration the upgrade from scheme 0
to 1.

This new scheme allows to add more metadata in the future.