Don't complain if the feed has an incorrect encoding.
feedparser has an Universal Parsing that will complain if the encoding
of the feed is incorrect (setting the bozo bit to 1) but it will then
try to detect the correct encoding.

The bozo_exception object will be CharacterEncodingOverride if the
initial encoding is wrong but feedparser was able to detect the correct
encoding anyways.

Make feed2maildir to complain only if feedparser could not detect the
encoding.